requirements.txt
dash
pandas
plotly
sqlalchemy
psycopg2-binary
openpyxl
gunicorn
python-dotenv